Renamed repository class and fixed up find and update methods
